grappling with TF 2 issues

+++ modules.py
@@ -79,8 +79,8 @@ def normalize(inputs,
         params_shape = inputs_shape[-1:]
 
         mean, variance = tf.nn.moments(x=inputs, axes=[-1], keepdims=True)
-        beta = tf.Variable(tf.zeros(params_shape))
-        gamma = tf.Variable(tf.ones(params_shape))
+        beta = tf.cast(tf.Variable(tf.zeros(params_shape)),dtype=tf.float64)
+        gamma = tf.cast(tf.Variable(tf.ones(params_shape)),dtype=tf.float64)
         normalized = (inputs - mean) / ((variance + epsilon) ** (.5))
         outputs = gamma * normalized + beta
